Rename questionType to question_type.dart
